Move pretraining head back to `examples/`
